requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bidict==0.22.1
blinker==1.7.0
click==8.1.7
cmake==3.27.7
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
exceptiongroup==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.104.1
Flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.44.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mysql-connector-python==8.2.0
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
annotated-types==0.6.0
anyio==3.7.1
bidict==0.22.1
blinker==1.7.0
click==8.1.7
cmake==3.27.7
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
exceptiongroup==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.104.1
Flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.44.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mysql-connector-python==8.2.0
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
annotated-types==0.6.0
anyio==3.7.1
bidict==0.22.1
blinker==1.7.0
click==8.1.7
cmake==3.27.7
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
exceptiongroup==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.104.1
Flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.44.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mysql-connector-python==8.2.0
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
annotated-types==0.6.0
anyio==3.7.1
bidict==0.22.1
blinker==1.7.0
click==8.1.7
cmake==3.27.7
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
exceptiongroup==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.104.1
Flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.44.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mysql-connector-python==8.2.0
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
